+++ src/main/kotlin/kravis/device/jupyter/JupyterIntegration.kt
@@ -1,13 +1,12 @@
-package kravis.jupyter
+package kravis.device.jupyter
 
 import kravis.GGPlot
-import org.jetbrains.kotlinx.jupyter.api.annotations.JupyterLibrary
 import org.jetbrains.kotlinx.jupyter.api.libraries.JupyterIntegration
 
 // reference https://github.com/Kotlin/kotlin-jupyter/blob/master/libraries/kravis.json
 
-@JupyterLibrary
-internal class Integration : JupyterIntegration() {
+@Suppress("unused")
+internal class JupyterIntegration : JupyterIntegration() {
     override fun Builder.onLoaded() {
         import("kravis.*")
         render<GGPlot> { it.show() }